requirements.txt
azure-storage-blob==12.12.0
azure-identity==1.10.0
Flask==2.0.2
python-dateutil==2.8.2
python-mimeparse==1.6.0
